requirements.txt
asgiref==3.2.10
astroid==2.4.2
asyncio==3.4.3
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
Django==3.1.2
django-appconf==1.0.4
django-cleanup==5.1.0
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.4.0
django-imagekit==4.0.2
django-jalali==4.0.0
django-jalali-date==0.3.1
django-phonenumber-field==5.0.0
djangorestframework==3.11.1
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.6.0
drf-nested-routers==0.92.5
drf-yasg==1.17.1
freezegun==1.0.0
idna==2.10
inflection==0.5.1
isort==5.6.4
itypes==1.2.0
jdatetime==3.6.2
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
model-bakery==1.2.1
mysql==0.0.2
mysqlclient==2.0.1
packaging==20.4
phonenumberslite==8.12.11
pilkit==2.0
Pillow==8.0.0
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.11
wrapt==1.12.1
